requirements.txt
certifi==2018.1.18
chardet==3.0.4
coverage==4.0.3
dj-database-url==0.4.2
Django==1.11
eventbrite==3.3.3
factory-boy==2.10.0
Faker==0.8.12
Flask==0.8
funcsigs==1.0.2
gunicorn==19.7.1
idna==2.6
ipaddress==1.0.19
Jinja2==2.6
mock==2.0.0
oauthlib==2.0.7
pbr==3.1.1
psycopg2-binary==2.7.4
PyJWT==1.6.1
python-coveralls==2.9.1
python-dateutil==2.7.1
python-openid==2.2.5
pytz==2018.3
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
text-unidecode==1.2
urllib3==1.22
Werkzeug==0.8.3
whitenoise==3.3.1
